PYIC-7876: Cleanup SSM
